v3: fix `make` bootstrap, and `[]int` stores into `[]i32` AST fields (#28408)

+++ vlib/v3/transform/transform.v
@@ -1222,7 +1222,7 @@ fn reserve_parallel_transform_ast_with_cache_mode(mut a flat.FlatAst, skip_gener
 	if grow_nodes {
 		old_nodes := a.nodes
 		a.nodes = []flat.Node{cap: nodes_cap}
-		a.file_node_ids = []int{}
+		a.file_node_ids = []i32{}
 		a.nodes << old_nodes
 	}
 	if grow_children {
@@ -3389,15 +3389,15 @@ fn (mut t Transformer) transform_serial_then_collect_pure(literal_decls []int) [
 	// ((previous top-level decl of ANY kind, fn_idx]); the shared-base parallel
 	// transform relies on those ranges being disjoint per item.
 	use_checker_tl := t.tc.top_level_idx.len > 0 && t.tc.top_level_idx_nodes_len == t.a.nodes.len
-	mut rebuilt_tl := []int{}
+	mut rebuilt_tl := []i32{}
 	if !use_checker_tl {
 		// Hand-built test ASTs and transforms after declaration synthesis do not
 		// have a current checker index. Rebuild only in that uncommon case.
-		rebuilt_tl = []int{cap: 1024}
+		rebuilt_tl = []i32{cap: 1024}
 		for i, node in t.a.nodes {
 			if node.kind in [.file, .module_decl, .struct_decl, .type_decl, .interface_decl,
 				.enum_decl, .import_decl, .const_decl, .global_decl, .fn_decl, .c_fn_decl] {
-				rebuilt_tl << i
+				rebuilt_tl << i32(i)
 			}
 		}
 	}
@@ -5007,7 +5007,7 @@ fn (mut t Transformer) transform_late_used_fn_bodies(names &[]string, names_star
 	late_scan_ids := if t.building_v && t.tc.top_level_idx.len > 0 {
 		t.tc.top_level_idx
 	} else {
-		[]int{}
+		[]i32{}
 	}
 	scan_count := if late_scan_ids.len > 0 { late_scan_ids.len } else { limit }
 	for scan_pos in 0 .. scan_count {
